requirements.txt

streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
gtts>=2.3.0
pyaudio>=0.2.11
streamlit-webrtc>=0.45.0
av>=10.0.0
numpy>=1.24.0